Added parseRemoteResponse(obj, resp) helper function
This function is called with the object and the response. It is only called
from dualsync() when the method is 'read' and no model was found in
localStorage. If the object has a method named 'parseBeforeLocalSave', response is run through that method. This allows for pre-parsing incoming JSON before it is saved locally, while also not interfering with the standard parse() method.

Also cleaned up resp, renaming it to 'response' throughout the script.